+++ src/org/yunghegel/gdx/utils/graphics/model/suppliers/CircleSupplier.java
@@ -36,7 +36,7 @@ public class CircleSupplier extends InstanceSupplier{
 
        v(0, centerY, 0);
         for (short i = 0; i < vertices; i++) {
-            short i1 = (short) ((short) i%vertices);
+            short i1 = (short) (i %vertices);
             short i2 = (short) ((short) (i+1)%vertices);
             b.triangle(i2, i1, (short) vertices);
         }
